Updated package.json for new Aurelia packages
Aurelia has updated their packages. One of the big updates is the updated dependency of core-js. Have two versions of core-js in a project slows down page load _significantly_. Because this package depended on old versions of Aurelia, it was pulling old versions of core-js, causing two versions of core-js to be loaded.